rename genericdecoder to decoder
