+++ dynamicstruct/decoder/json_decoder.go
@@ -1,21 +1,21 @@
-package genericdecoder
+package decoder
 
 import (
 	"encoding/json"
 )
 
-// JSONGenericDecoder is the decoded object from unknown format JSON/YAML/and others.
-type JSONGenericDecoder struct {
+// JSONDecoder is the decoded object from unknown format JSON/YAML/and others.
+type JSONDecoder struct {
 }
 
-// NewJSONGenericDecoder returns a concrete GenericDecoder for JSON
-func NewJSONGenericDecoder() GenericDecoder {
-	return &JSONGenericDecoder{}
+// NewJSONDecoder returns a concrete Decoder for JSON
+func NewJSONDecoder() Decoder {
+	return &JSONDecoder{}
 }
 
 // Decode decodes JSON data to interface via DynamicStruct.DecodeMap.
 // data argument must be a byte array data of valid JSON.
-func (jgd *JSONGenericDecoder) Decode(data []byte) (*DecodedResult, error) {
+func (jgd *JSONDecoder) Decode(data []byte) (*DecodedResult, error) {
 	// FIXME:
 	// want to add json validation. but is json.Valid(data) too slow?
 	// See: https://stackoverflow.com/questions/22128282/how-to-check-string-is-in-json-format
